requirements-dev.txt
-r requirements.txt

model-mommy==1.2.5
pytest==2.7.2
pytest-django==2.8.0
